% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTestFunction.R
\name{plot.hzhf_function}
\alias{plot.hzhf_function}
\title{Plot function for hzhf function}
\usage{
\method{plot}{hzhf_function}(x, resolution = 128, ids = NULL,
  def.vec = 0.35 * getZMax(f), ...)
}
\arguments{
\item{x}{[\code{function}] \cr
Size of parameter space.}

\item{resolution}{[\code{integer(1)}] \cr
Number non-hierarchical parameters. Must be small than in.dim - 1.}

\item{ids}{[\code{double(1)}] \cr
Metaparameter of all test functions, size of hierarchical area, see details. Default is 0.5.}

\item{def.vec}{[\code{double(1)}] \cr
Metaparameter of all test functions, shift of hierarchical area, see details.. Default is 0.}

\item{...}{[any] \cr
Not used.}
}
\value{
A \code{smoof_function}.
}
\description{
Plot function for hzhf function
}
