% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hzhf.R, R/hzhf01.R, R/hzhf02.R, R/hzhf03.R,
%   R/hzhf04.R, R/hzhf05.R, R/hzhf06.R, R/hzhf07.R, R/hzhf08.R, R/hzhf09.R,
%   R/hzhf10.R, R/hzhf11.R, R/hzhf12.R, R/hzhf13.R, R/hzhf14.R, R/hzhf15.R,
%   R/hzhf_doku.R
\name{makeHZHF}
\alias{makeHZHF}
\alias{makeHZHF01}
\alias{makeHZHF02}
\alias{makeHZHF03}
\alias{makeHZHF04}
\alias{makeHZHF05}
\alias{makeHZHF06}
\alias{makeHZHF07}
\alias{makeHZHF08}
\alias{makeHZHF09}
\alias{makeHZHF10}
\alias{makeHZHF11}
\alias{makeHZHF12}
\alias{makeHZHF13}
\alias{makeHZHF14}
\alias{makeHZHF15}
\alias{hzhf}
\alias{HZHF}
\title{Generators for HZHF test function}
\usage{
makeHZHF(id, in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF01(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF02(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF03(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF04(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF05(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF06(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF07(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF08(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF09(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF10(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF11(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF12(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF13(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF14(in.dim, k, c = 0.5, s = 0, check = TRUE)

makeHZHF15(in.dim, k, c = 0.5, s = 0, check = TRUE)
}
\arguments{
\item{id}{[\code{integer(1)}] \cr
ID of hzhf function to generate, element of {1, 2, ..., 15}.}

\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{k}{[\code{integer(1)}] \cr
Number non-hierarchical parameters. Must be small than in.dim - 1.}

\item{c}{[\code{double(1)}] \cr
Metaparameter of all test functions, size of hierarchical area, see details. Default is 0.5.}

\item{s}{[\code{double(1)}] \cr
Metaparameter of all test functions, shift of hierarchical area, see details.. Default is 0.}

\item{check}{[\code{logical(1)}] \cr
Shall parameter checks be performed befor each function evaluation? Default is TRUE}
}
\value{
A \code{hzhf_function}, subclass of \code{smoof_function}.
}
\description{
Generators for HZHF test function
}
